requirements.txt
streamlit==1.26.1
openai==0.27.8
pandas==2.3.1
python-dotenv==1.0.0
tiktoken>=0.4.2
